Move generic lambda logic in a common super type
